backend/requirements.txt
Django
djangorestframework
graphene-django
django-cors-headers
djangorestframework_simplejwt
django-allauth
dj-rest-auth
django-filter
channels
channels-redis
PyJWT
pillow


